+++ zcash_keys/CHANGELOG.md
@@ -8,6 +8,7 @@ and this library adheres to Rust's notion of
 ### Added
 - `zcash_keys::address::Address::try_from_zcash_address`
 - `zcash_keys::address::Receiver`
+- `zcash_keys::keys::UnifiedAddressRequest::intersect`
 
 ### Changed
 - MSRV is now 1.70.0.